+++ .gitignore
@@ -1,5 +1,5 @@
 # MINE
-# build dir
+bin/
 build/
 
 # Log output
