+++ src/tools/work_factor_computation.cpp
