+++ src/tools/work_factor_computation_parallel.cpp
